ml-service/requirements.txt
--extra-index-url https://download.pytorch.org/whl/cpu

fastapi==0.115.0
uvicorn[standard]==0.32.0
sentence-transformers==3.2.1
scikit-learn==1.5.2
numpy==1.26.4
pydantic==2.9.2
slowapi==0.1.9
python-dotenv==1.0.1